requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-countries==7.2.1
django-environ==0.8.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dropbox==11.22.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
ply==3.11
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
stone==3.2.1
urllib3==1.26.7
whitenoise==5.3.0
